- organized unit tests - removed jsdom-global from Spectron base file (brave.js), moved to a unit-test only file (braveUnit.js) - updated tests.md - added jsdom-global to ignoredPaths.js
