+++ test/unit/braveUnit.js
@@ -1,0 +1,1 @@
+require('jsdom-global')()
